chore: refator file dashboard to page
