+++ web/src/less/resources-dialog.less
@@ -58,91 +58,6 @@
           }
         }
 
-        .resource-card{
-          @apply flex flex-col relative justify-start hover:bg-slate-200 w-48 m-4 rounded-md	;
-
-          &:hover {
-            .resource-checkbox {
-              display: flex;
-      
-            }
-            .more-action-btn{
-              display: flex;
-            }
-          }
-          .resource-checkbox-selected{
-            /* position: absolute; */
-          }
-
-          .resource-checkbox{
-            @apply hidden hover:flex;
-          }
-
-
-          .more-action-btn{
-            @apply hidden;
-            &:hover {
-              & + .more-action-btns-wrapper {
-                display: flex;
-              }
-            }  
-          }
-              
-          .icon-cover{
-            @apply w-32 h-32 ml-auto mr-auto mt-3;
-          }
-
-
-          .resource-title{
-            @apply text-base overflow-ellipsis text-center	;
-          }
-
-          .resource-time{
-            @apply text-sm text-gray-400 text-center	;
-          }
-
-          .btns-container{
-            @apply w-full flex flex-row-reverse justify-between  h-1; // the h-1 is must, making the checkbox and more don't move when hover
-            > .more-action-btns-wrapper {
-              @apply hidden flex-col justify-start items-center absolute top-2 -right-4 flex-nowrap hover:flex p-3;
-      
-              > .more-action-btns-container {
-                @apply w-auto h-auto flex flex-col p-1 z-1 whitespace-nowrap rounded-lg bg-white dark:bg-zinc-800;
-                box-shadow: 0 0 8px 0 rgb(0 0 0 / 20%);
-      
-                > .btns-container {
-                  @apply w-full flex flex-row justify-around items-center border-b border-gray-100 dark:border-zinc-600 p-1 mb-1;
-      
-                  > .btn {
-                    @apply relative w-6 h-6 p-1 text-gray-600 dark:text-gray-300 cursor-pointer select-none;
-      
-                    &:hover > .tip-text {
-                      @apply block;
-                    }
-      
-                    > .icon-img {
-                      @apply w-5 h-auto;
-                    }
-      
-                    > .tip-text {
-                      @apply hidden absolute top-0 p-1 px-2 rounded text-xs leading-6 -mt-9 bg-black text-white shadow opacity-70;
-                    }
-                  }
-                }
-      
-                > .btn {
-                  @apply w-full text-sm leading-6 py-1 px-3 rounded justify-start cursor-pointer dark:text-gray-300;
-      
-                  &.archive-btn {
-                    @apply text-orange-600;
-                  }
-                }
-              }
-            }
-      
-            
-          }
-        }
 
         .field-text {
           @apply w-full truncate text-base pr-2 last:pr-0;
